print integer to double

+++ Chap_01/fahr_to_celsius.c
@@ -7,14 +7,14 @@
 
 #define DEGREE_TABLE_MAX		16
 
-int	fahr_to_celsius(int fahr)
+double	fahr_to_celsius(double fahr)
 {
-	return ((int)(((double)fahr - 32) * 5.0 / 9.0));
+	return ((int)((fahr - 32) * 5.0 / 9.0));
 }
 
-int	main(void)
+int		main(void)
 {
-	int	degree[DEGREE_TABLE_MAX][2] = {
+	double	degree[DEGREE_TABLE_MAX][2] = {
 		{0, -17},
 		{20, -6},
 		{40, 4},
@@ -33,7 +33,7 @@ int	main(void)
 		{300, 148}};
 	
 	for (int idx = 0; idx < DEGREE_TABLE_MAX; idx++)
-		printf("fahr: %3d, fahr_to_celsius(): %3d, celsius: %3d\n"
-				, degree[idx][0], fahr_to_celsius(degree[idx][0]), degree[idx][1]);
+		printf("fahr: %3.0f, fahr_to_celsius(): %5.1f, celsius: %5.1f\n"
+				, degree[idx][0], fahr_to_celsius((double)degree[idx][0]), degree[idx][1]);
 	return (0);
 }
